% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{._analyzeHeaderFormat}
\alias{._analyzeHeaderFormat}
\title{Analyze FASTQ header format to determine platform characteristics}
\usage{
._analyzeHeaderFormat(headers)
}
\arguments{
\item{headers}{Character vector of FASTQ headers}
}
\value{
Character string describing the header format
}
\description{
Analyze FASTQ header format to determine platform characteristics
}
\keyword{internal}
