% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{._createDirIfNotExists}
\alias{._createDirIfNotExists}
\title{Create a directory if it doesn't exist}
\usage{
._createDirIfNotExists(dir_path)
}
\arguments{
\item{dir_path}{Character string specifying directory path}
}
\value{
Logical indicating if the directory exists or was created
}
\description{
Create a directory if it doesn't exist
}
\keyword{internal}
