% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware_detection.R
\name{._detectArchitecture}
\alias{._detectArchitecture}
\title{Detect CPU architecture}
\usage{
._detectArchitecture()
}
\value{
Character string with CPU architecture
}
\description{
Detect CPU architecture
}
\keyword{internal}
