% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware_detection.R
\name{._detectPhysicalCores}
\alias{._detectPhysicalCores}
\title{Detect physical CPU cores}
\usage{
._detectPhysicalCores()
}
\value{
Integer number of physical CPU cores
}
\description{
Detect physical CPU cores
}
\keyword{internal}
