% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{._detectPotentialFileType}
\alias{._detectPotentialFileType}
\title{Detect whether a file matches forward or reverse patterns}
\usage{
._detectPotentialFileType(file_names, patterns)
}
\arguments{
\item{file_names}{Character vector of file names}

\item{patterns}{Character vector of patterns to match}
}
\value{
Logical vector indicating matches
}
\description{
Detect whether a file matches forward or reverse patterns
}
\keyword{internal}
