% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{._dirExists}
\alias{._dirExists}
\title{Check if a directory exists}
\usage{
._dirExists(dir_path)
}
\arguments{
\item{dir_path}{Character string specifying directory path}
}
\value{
Logical indicating if the directory exists
}
\description{
Check if a directory exists
}
\keyword{internal}
