% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{._endsWith}
\alias{._endsWith}
\title{Check if a string ends with a pattern}
\usage{
._endsWith(x, pattern)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Character string with pattern to match at end}
}
\value{
Logical vector indicating which elements end with pattern
}
\description{
Check if a string ends with a pattern
}
\keyword{internal}
