% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{._extractMetadataFromFilename}
\alias{._extractMetadataFromFilename}
\title{Extract metadata from a filename using regular expressions}
\usage{
._extractMetadataFromFilename(filename)
}
\arguments{
\item{filename}{Character string of filename}
}
\value{
List containing extracted metadata
}
\description{
Extract metadata from a filename using regular expressions
}
\keyword{internal}
