% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{._fileExists}
\alias{._fileExists}
\title{Validate that a file exists}
\usage{
._fileExists(file_path)
}
\arguments{
\item{file_path}{Character string specifying file path}
}
\value{
Logical indicating if the file exists
}
\description{
Validate that a file exists
}
\keyword{internal}
