% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{optimizeFilteringParams}
\alias{optimizeFilteringParams}
\title{Optimize filtering parameters}
\usage{
optimizeFilteringParams(
  fastq_files,
  output_dir,
  truncLen_range = list(c(240, 200), c(220, 180), c(200, 160)),
  maxEE_range = list(c(2, 2), c(2, 5), c(5, 5)),
  sample_size = 2,
  paired = TRUE
)
}
\arguments{
\item{fastq_files}{Tibble with FASTQ file information as produced by detectFastqFiles()}

\item{output_dir}{Character string specifying temporary output directory for filtered files}

\item{truncLen_range}{List of integer vectors to test for truncation lengths}

\item{maxEE_range}{List of double vectors to test for maximum expected errors}

\item{sample_size}{Integer indicating number of samples to use for optimization}

\item{paired}{Logical indicating whether files are paired}
}
\value{
A tibble with filtering results for different parameter combinations
}
\description{
This function helps identify optimal filtering parameters by testing different
combinations of truncation lengths and maxEE values.
}
\examples{
\dontrun{
fastq_files <- detectFastqFiles("path/to/fastq/dir")
truncLen_range <- list(c(240, 200), c(220, 180), c(200, 160))
maxEE_range <- list(c(2, 2), c(2, 5), c(5, 5))
opt_params <- optimizeFilteringParams(fastq_files, "path/to/temp/dir",
                                     truncLen_range, maxEE_range)
}
}
